Project Abstract
The Project HOTEL MANAGEMENT
SYSTEM is an application written in the C programming language that allows the hotel manager to handle
all hotel activities on a computer. Its User Interface and
the ability to manage various hotel bookings and rooms
make this system very flexible and convenient. The
hotel manager is a very busy person and does not have
the time to sit and manage the entire activities
manually on paper. This application gives him/her the
power and flexibility to manage the entire system from a
single text file. The Hotel Management project provides
room booking, availability, cancellation and other
necessary hotel management tools like features of the
room, customer details, Restaurant, and invoices. The
system allows the manager to post room availability status in the system. Customers can also cancel their booking/bookings. The customers can also view the restaurant menus and other
hotel services. They can also order food, and drinks they
want and get the bill
according to the quantity of their order. The system is
hence useful for both customers and managers to portably
manage hotel activities